+++ differt/src/differt/plugins/__init__.py
@@ -1,0 +1,4 @@
+"""DiffeRT plugins.
+
+Plugins extend the functionality of DiffeRT by providing additional features or integrations with other libraries.
+"""